fix(file-records): file_number bigint overflow guard + real date validation

+++ src/Domain/Entities/FileRecord.cs
@@ -10,5 +10,5 @@ public class FileRecord
     public int? FlopDiskNumber { get; set; }
     public DateTime Date { get; set; }
     public string Client { get; set; } = string.Empty;
-    public int FileNumber { get; set; }
+    public long FileNumber { get; set; }
 }
